backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==44.0.0

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Task Queue
celery==5.4.0
redis==5.2.0
flower==2.0.1

# AWS
boto3==1.35.82
botocore==1.35.82

# Plaid
plaid-python==29.0.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
faker==33.1.0

# Code Quality
ruff==0.8.4
mypy==1.13.0
black==24.10.0
